requirement.txt
python-decouple==3.6
azure-cognitiveservices-speech==1.20.0
bcrypt==3.1.7
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Jinja2==2.11.3
kanpai==0.1.15
Pillow==7.0.0
pymongo==4.0.1
python-dotenv==0.19.2
requests==2.22.0
MarkupSafe==1.1.1
itsdangerous==1.1.0
Werkzeug==0.16.0
